requirements.txt
faster-whisper
pyaudio
mss
Pillow
requests
numpy
pystray
pywin32
torch
pypdf
python-docx
